feat: bring your own logger
Pino is no longer a dependency. We now require that you provide your own logger (such as... pino!) if you'd like logging.

This is similar to what I did for @waveplay/pilot. Doing it this way not only lowers the bundle size, but it also makes it easier to have advanced logger functionality by providing more flexible control.