fix: declare npm devDependency explicitly
This is kind of a weird issue. Sub packages trying to use anything that relies on `npm` accidentally find this package which is installed by semantic-release and execute it instead.

This is a problem when `npm exec` is run, which the version that semantic release bundles doesn't support and Pilot.js requires for the `dev` command in the sample project...